test: split worker runtime tests and bound startup waits
