+++ tasks/lessons.md
@@ -161,6 +161,7 @@
 - If timeout or recovery code publishes in-memory task updates, every producer and waiter must share the same `TaskSchedulerService` instance or clones of it. Reconstructing a fresh service with the same repositories splits `task_waiters` state and breaks notifications.
 - If a Mongo/task mapper validates retry invariants when reading documents back, mirror that validation at the write boundary too so invalid retry strategies cannot be persisted as poison rows.
 - Process-scoped `OnceLock` temp fixtures with deterministic paths should proactively clear stale directories on initialization because `Drop` cleanup will not run at test-binary exit.
+- When a scheduler or worker test waits for `Notify`, `watch`, or channel startup signals, never block on the raw await without a timeout. Wrap the wait in `tokio::time::timeout(...)` or a helper so broken startup synchronization fails fast instead of hanging the suite.
 
 - For websocket-backed LLM chats, a single initial typing/progress frame is not enough when the backend can spend tens of seconds waiting on the provider. If the handler then blocks silently on the long-running reply, proxies can drop the socket even though the server eventually persists the final answer. Use a keepalive loop that emits periodic progress frames until the reply task completes.
 - If a test exists only to verify timer-driven behavior like websocket keepalives or delayed retries, do not make it wait on real wall-clock time. Use paused Tokio time (`#[tokio::test(start_paused = true)]`) and explicit `tokio::time::advance(...)` so the test is fast and deterministic.
